# ⚡️ Only load relation's kf_id