@dankolbman 2018-04-04
Yeah, this pr may not be relevant anymore for the aliquot+samples. We should go back and inspect all of the entities with fks, though, which is just about everything. I will close this, but we can still use it for reference.